OPHJOD-450: Refactor routes to use outlet
