+++ src/main.tsx
@@ -12,11 +12,9 @@ const container = document.getElementById('root')!;
 const root = createRoot(container);
 
 if (process.env.NODE_ENV !== 'production') {
-  /* eslint-disable @typescript-eslint/no-floating-promises */
-  import('@axe-core/react').then((axe) => {
-    axe.default(React, root, 1000);
+  void import('@axe-core/react').then((axe) => {
+    void axe.default(React, root, 1000);
   });
-  /* eslint-enable @typescript-eslint/no-floating-promises */
 
   onCLS(console.log);
   onFID(console.log);
