+++ src/routes/BasicInformation/index.ts
@@ -1,0 +1,8 @@
+import BasicInformation from './BasicInformation';
+import CookiePolicy from './CookiePolicy';
+import DataSources from './DataSources';
+import TermsOfService from './TermsOfService';
+import AccessibilityStatement from './AccessibilityStatement';
+import PrivacyPolicy from './PrivacyPolicy';
+
+export { BasicInformation, CookiePolicy, DataSources, TermsOfService, AccessibilityStatement, PrivacyPolicy };
