+++ src/routes/Profile/index.ts
@@ -1,0 +1,19 @@
+import Profile from './Profile';
+import Preferences from './Preferences';
+import Favorites from './Favorites';
+import Competences from './Competences';
+import WorkHistory from './WorkHistory';
+import EducationHistory from './EducationHistory';
+import FreeTimeActivities from './FreeTimeActivities';
+import SomethingElse from './SomethingElse';
+
+export {
+  Profile,
+  Preferences,
+  Favorites,
+  Competences,
+  WorkHistory,
+  EducationHistory,
+  FreeTimeActivities,
+  SomethingElse,
+};
