+++ src/routes/Root/Root.tsx
@@ -1,17 +1,11 @@
 import { useRef, type ComponentProps } from 'react';
-import { Route, Routes, NavLink, useLocation, matchPath, ScrollRestoration, useLoaderData } from 'react-router-dom';
+import { NavLink, useLocation, matchPath, ScrollRestoration, useLoaderData, Outlet } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { Helmet } from 'react-helmet-async';
 import { ActionBarContext } from '@/hooks/useActionBar';
 import { ErrorNote } from '@/features';
 import { NavigationBar, Footer, SkipLink } from '@jod/design-system';
 import { RootLoaderData } from './loader';
-import Home from '@/routes/Home';
-import UserGuide from '@/routes/UserGuide';
-import BasicInformation from '@/routes/BasicInformation';
-import Tool from '@/routes/Tool';
-import PersonalPages from '@/routes/PersonalPages';
-import NoMatch from '@/routes/NoMatch';
 import { AuthContext } from '@/hooks/useAuth';
 
 const NavigationBarItem = (to: string, text: string) => ({
@@ -26,10 +20,10 @@ const NavigationBarItem = (to: string, text: string) => ({
 const Root = () => {
   const { t, i18n } = useTranslation();
   const { pathname } = useLocation();
-  const userGuide = `/${i18n.language}/${t('slugs.user-guide')}`;
+  const userGuide = `/${i18n.language}/${t('slugs.user-guide.index')}`;
   const basicInformation = `/${i18n.language}/${t('slugs.basic-information')}`;
   const items: ComponentProps<typeof Footer>['items'] = [
-    NavigationBarItem(`${userGuide}/${t('slugs.what-is-the-service')}`, t('about-us-and-user-guide')),
+    NavigationBarItem(`${userGuide}/${t('slugs.user-guide.what-is-the-service')}`, t('about-us-and-user-guide')),
     NavigationBarItem(`${basicInformation}/${t('slugs.cookie-policy')}`, t('cookie-policy')),
     NavigationBarItem(`${basicInformation}/${t('slugs.data-sources')}`, t('data-sources')),
     NavigationBarItem(`${basicInformation}/${t('slugs.terms-of-service')}`, t('terms-of-service')),
@@ -46,7 +40,7 @@ const Root = () => {
   }));
   // If homepage, use light variant, otherwise use dark variant
   const variant: ComponentProps<typeof Footer>['variant'] = matchPath(`/${i18n.language}`, pathname) ? 'light' : 'dark';
-  const ref = useRef<HTMLDivElement>(null);
+  const footerRef = useRef<HTMLDivElement>(null);
 
   const data = useLoaderData() as RootLoaderData;
 
@@ -71,10 +65,7 @@ const Root = () => {
               name: 'Reetta Räppänä',
               component: ({ children, ...rootProps }) => {
                 return (
-                  <NavLink
-                    to={`/${i18n.language}/${t('slugs.personal-pages.index')}/${t('slugs.personal-pages.preferences')}`}
-                    {...rootProps}
-                  >
+                  <NavLink to={`/${i18n.language}/${t('slugs.profile.index')}`} {...rootProps}>
                     {children}
                   </NavLink>
                 );
@@ -85,17 +76,10 @@ const Root = () => {
         />
         <ErrorNote />
       </header>
-      <ActionBarContext.Provider value={ref.current}>
-        <Routes>
-          <Route index element={<Home />} />
-          <Route path={`${t('slugs.tool.index')}/*`} element={<Tool />} />
-          <Route path={`${t('slugs.personal-pages.index')}/*`} element={<PersonalPages />} />
-          <Route path={`${t('slugs.user-guide')}/*`} element={<UserGuide />} />
-          <Route path={`${t('slugs.basic-information')}/*`} element={<BasicInformation />} />
-          <Route path="*" element={<NoMatch />} />
-        </Routes>
+      <ActionBarContext.Provider value={footerRef.current}>
+        <Outlet />
       </ActionBarContext.Provider>
-      <Footer ref={ref} items={items} logos={logos} copyright={t('copyright')} variant={variant} />
+      <Footer ref={footerRef} items={items} logos={logos} copyright={t('copyright')} variant={variant} />
     </AuthContext.Provider>
   );
 };
