+++ src/routes/Root/index.ts
@@ -1,6 +1,6 @@
-import Root from './Root';
 import loader from './loader';
+import Root from './Root';
+import ErrorElement from './ErrorElement';
+import NoMatch from './NoMatch';
 
-export { loader };
-
-export default Root;
+export { loader, Root, ErrorElement, NoMatch };
