+++ src/routes/Tool/index.ts
@@ -1,0 +1,9 @@
+import Tool from './Tool';
+import Instructions from './Instructions';
+import Goals from './Goals';
+import Competences from './Competences';
+import Interests from './Interests';
+import Restrictions from './Restrictions';
+import Search from './Search';
+
+export { Tool, Instructions, Goals, Competences, Interests, Restrictions, Search };
