+++ src/routes/UserGuide/WhatIsTheService.tsx
@@ -1,0 +1,23 @@
+import { useTranslation } from 'react-i18next';
+import { Title } from '@/components';
+
+const WhatIsTheService = () => {
+  const { t } = useTranslation();
+  const title = t('what-is-the-service');
+
+  return (
+    <>
+      <Title value={title} />
+      <h1 className="mb-5 text-heading-2 sm:text-heading-1">{title}</h1>
+      <p className="mb-8 text-body-md">
+        Non curabitur gravida arcu ac tortor dignissim. Eget lorem dolor sed viverra ipsum. Nibh mauris cursus mattis
+        molestie a iaculis at erat pellentesque. Sagittis orci a scelerisque purus. Consequat nisl vel pretium lectus
+        quam id leo in. Id faucibus nisl tincidunt eget nullam non nisi. Nullam vehicula ipsum a arcu cursus. Arcu ac
+        tortor dignissim convallis aenean et. Pellentesque pulvinar pellentesque habitant morbi tristique. Pharetra et
+        ultrices neque ornare aenean euismod elementum nisi.
+      </p>
+    </>
+  );
+};
+
+export default WhatIsTheService;
