+++ src/routes/UserGuide/index.ts
@@ -1,0 +1,17 @@
+import UserGuide from './UserGuide';
+import WhatIsTheService from './WhatIsTheService';
+import WhoIsTheServiceFor from './WhoIsTheServiceFor';
+import HowDoIUseTheService from './HowDoIUseTheService';
+import WhereCanIGetMoreHelp from './WhereCanIGetMoreHelp';
+import WhoProvidesTheService from './WhoProvidesTheService';
+import HowDoIGiveFeedback from './HowDoIGiveFeedback';
+
+export {
+  UserGuide,
+  WhatIsTheService,
+  WhoIsTheServiceFor,
+  HowDoIUseTheService,
+  WhereCanIGetMoreHelp,
+  WhoProvidesTheService,
+  HowDoIGiveFeedback,
+};
